PIN En/Decrypt, changelog and readme documentation
Possibly final push. Implemented PIN-based encryption and decryption in the EncryptionService.java. Removed ToDoList, and added the changelog and readme. Instructions on how to use the EncryptionService are detailed in the readme. Also cleaned up the ENCRYPT and DECRYPT functions used for files, and added a necessary string input to DECRYPT for the user to define where to store the decrypted file.